requirements.txt
streamlit
ollama
chromadb
langchain-community
langchain-core
langchain-text-splitters
sentence-transformers
python-dotenv
pymupdf
python-docx